Revert "clean project structure"
