+++ settings.json
@@ -1,10 +1,18 @@
 {
-  "city": "Gafsa",
-  "country": "Tunisia",
+  "city": "Quebec",
+  "country": "Canada",
   "theme": "green",
   "language": "fr",
   "position": {
-    "x": 469,
-    "y": 121
+    "x": 150,
+    "y": 144
+  },
+  "adhanEnabled": true,
+  "adhanEnabledByPrayer": {
+    "Fajr": true,
+    "Dhuhr": true,
+    "Asr": true,
+    "Maghrib": true,
+    "Isha": true
   }
 }
